Set up uWSGI and nginx web servers for remote dev environment
wsgi.py enables the uWSGI web server to initialize the Flask app. The
conditional statements in wsgi.py and hello.py ensure that only one
instance of the app is run, and exposes that instance on
localhost:5001.

Updates to requirements.txt reflect the addition of the uwsgi package.

appserver.ini contains various options passed to uwsgi on starting the
Flask app and interacting with it via multiple threads. Also creates a
socket for access by the nginx reverse proxy.

nginx instance on the development machine requires further configs in
order to connect http traffic with the uWSGI server, but this is
beyond the scope of this commit. Specific instructions for these
configs will be added to README, but see https://hackernoon.com/a-
guide-to-scaling-machine-learning-models-in-production-aa8831163846
in the meantime for a similar setup process.