Add web form support with Flask-WTF extension
Flask-WTF provides helpers that allow Python to interact with the
WTForm.js library. Flask-Bootstrap contains a handy base template
that applies nice styles to WTForm objects.

A simple test form was added to the root page of the app to
demonstrate field validation and data submission via POST request.

NOTE: The SECRET_KEY variable included is for learning purposes and
convenience, but should never be used for an internet-facing
deployment. Remove it from the source and recreate it as a randomly
generated environment variable before deploymnet.