Add shell context processor
Shell context processor programmatically imports certain objects for use
in the Flask shell. Adding objects to the processor's arguments makes it
convenient to access these objects when interacting with the app via the
shell interface, as the objects do not need to be re-imported each time
a new shell session is opened.